refactor(cockpit): move risk tests to owner module
